+++ R/score_recurrence.R
@@ -1,0 +1,54 @@
+#' @title Score signals based on temporal alarm recurrence within strata
+#' @description Computes a score for each signal based on the frequency of alarms
+#' within the test period of its stratum and category. Higher alarm frequency results in higher scores.
+#' The score is set to `NA` when recurrence cannot be assessed, such as when the test period consists of only one week.
+#' A single alarm across multiple observed weeks is assigned a score of 0, because it does not indicate recurrence.
+#' If more than one alarm occurs, the score is calculated as `n_alarms / n_weeks`.
+#'
+#' @param signal_results signal detection results obtained from get_signals()
+#'
+#' @returns signal detection results scored
+#' @export
+#'
+#' @examples \dontrun{
+#' signals <- input_example %>%
+#'   preprocess_data() %>%
+#'   get_signals_all() %>%
+#'   dplyr::mutate(pathogen = "Pertussis")
+#'
+#' get_scores(signals, scorers = list(recurrence = score_recurrence))
+#' }
+score_recurrence <- function(signal_results) {
+  # select test period per stratum
+  ts_pathogen_strat <- signal_results %>%
+    dplyr::group_by(.data$category, .data$stratum) %>%
+    dplyr::arrange(.data$year, .data$week, .by_group = TRUE) %>%
+    dplyr::group_modify(~ {
+      n <- dplyr::first(.x$number_of_weeks)
+      dplyr::slice_tail(.x, n = n)
+    }) %>%
+    dplyr::ungroup()
+
+  # count alarms and compute individual score for each stratum
+  score_stratum <- ts_pathogen_strat %>%
+    dplyr::group_by(.data$category, .data$stratum) %>%
+    dplyr::summarise(
+      n_weeks = dplyr::n(),
+      n_alarms = sum(.data$alarms, na.rm = TRUE),
+      score_stratum = dplyr::if_else(n_alarms > 0, n_alarms / n_weeks, NA),
+      score_stratum = dplyr::if_else(n_alarms == 1, 0, score_stratum),
+      score_stratum = dplyr::if_else(n_weeks == 1, NA, score_stratum)
+    ) %>%
+    dplyr::ungroup()
+
+  # add row_id
+  score_tbl <- signal_results %>%
+    dplyr::select(".row_id", "category", "stratum", "alarms") %>%
+    dplyr::left_join(score_stratum, by = c("category", "stratum")) %>%
+    dplyr::mutate(
+      score = dplyr::if_else(.data$alarms %in% TRUE, .data$score_stratum, NA)
+    ) %>%
+    dplyr::select(".row_id", "score")
+
+  return(score_tbl)
+}